+++ specifications/src/main/java/org/dcsa/conformance/specifications/generator/StandardSpecification.java
@@ -22,6 +22,7 @@ import lombok.extern.slf4j.Slf4j;
 import org.dcsa.conformance.specifications.constraints.SchemaConstraint;
 import org.dcsa.conformance.specifications.dataoverview.DataOverview;
 import org.dcsa.conformance.specifications.dataoverview.DataOverviewSheet;
+import org.dcsa.conformance.specifications.dataoverview.LegendMetadata;
 
 @Slf4j
 public abstract class StandardSpecification {
@@ -90,7 +91,17 @@ public abstract class StandardSpecification {
                                             .add(schemaConstraint))));
 
     ModelConverters.getInstance()
-        .addConverter(new ModelValidatorConverter(constraintsByClassAndField));
+        .getConverters()
+        .forEach(
+            existingConverter -> {
+              if (existingConverter instanceof ModelValidatorConverter) {
+                log.info("Removing {}", existingConverter);
+                ModelConverters.getInstance().removeConverter(existingConverter);
+              }
+            });
+    ModelConverters.getInstance()
+        .addConverter(
+            new ModelValidatorConverter(constraintsByClassAndField, modelClassesStream()));
     modelClassesStream()
         .sorted(Comparator.comparing(Class::getSimpleName))
         .forEach(
@@ -99,6 +110,8 @@ public abstract class StandardSpecification {
                     .read(modelClass)
                     .forEach(openAPI.getComponents()::addSchemas));
   }
+
+  protected abstract LegendMetadata getLegendMetadata();
 
   protected abstract Stream<Class<?>> modelClassesStream();
 
@@ -128,6 +141,7 @@ public abstract class StandardSpecification {
 
     DataOverview dataOverview =
         new DataOverview(
+            getLegendMetadata(),
             constraintsByClassAndField,
             SpecificationToolkit.parameterizeStringRawSchemaMap(
                 openAPI.getComponents().getSchemas()),