+++ specifications/src/main/java/org/dcsa/conformance/specifications/standards/an/v100/model/VesselVoyage.java
@@ -2,7 +2,6 @@ package org.dcsa.conformance.specifications.standards.an.v100.model;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.Data;
-import org.dcsa.conformance.specifications.generator.SchemaOverride;
 import org.dcsa.conformance.specifications.standards.an.v100.types.CountryCode;
 import org.dcsa.conformance.specifications.standards.an.v100.types.UniversalVoyageReference;
 import org.dcsa.conformance.specifications.standards.an.v100.types.VesselIMONumber;
@@ -26,23 +25,21 @@ the first mother vessel arriving at the destination country.
 """)
 public class VesselVoyage {
 
-  @Schema(requiredMode = Schema.RequiredMode.REQUIRED)
+  @Schema()
   private VesselVoyageTypeCode typeCode;
 
   @Schema(
-      requiredMode = Schema.RequiredMode.REQUIRED,
       maxLength = 50,
       example = "King of the Seas",
       description = "Vessel name")
   private String vesselName;
 
-  @SchemaOverride(description = "Vessel flag")
+  @Schema(description = "Vessel flag")
   private CountryCode vesselFlag;
 
   @Schema() private VesselIMONumber vesselIMONumber;
 
   @Schema(
-      requiredMode = Schema.RequiredMode.REQUIRED,
       maxLength = 50,
       example = "1234N",
       description =
@@ -53,7 +50,7 @@ Carrier-specific identifier of a voyage:
 """)
   private String carrierVoyageNumber;
 
-  @SchemaOverride(
+  @Schema(
       description =
 """
 Universal identifier of a voyage: