+++ specifications/src/main/java/org/dcsa/conformance/specifications/standards/an/v100/types/FormattedDate.java
@@ -6,7 +6,7 @@ import lombok.Data;
 @Data
 @Schema(
     type = "string",
-    pattern = "^\\d{4}-\\d{2}-\\d{2}$",
+    format = "date",
     example = "2025-01-23",
     description = "String representation of a date in yyyy-MM-dd format.")
 public class FormattedDate {}