+++ specifications/src/main/java/org/dcsa/conformance/specifications/standards/dt/v100/model/ConsignmentItem.java
@@ -4,36 +4,36 @@ import io.swagger.v3.oas.annotations.media.ArraySchema;
 import io.swagger.v3.oas.annotations.media.Schema;
 import java.util.List;
 import lombok.Data;
+import org.dcsa.conformance.specifications.standards.dt.v100.types.DescriptionOfGoodsLine;
+import org.dcsa.conformance.specifications.standards.dt.v100.types.HSCode;
+import org.dcsa.conformance.specifications.standards.dt.v100.types.ShippingMark;
 
-@Schema(description = "Defines a list of `CargoItems` belonging together and the associated `Booking`. A `ConsignmentItem` can be split across multiple containers by referencing multiple `CargoItems`.")
+@Schema(description = ConsignmentItem.CLASS_SCHEMA_DESCRIPTION)
 @Data
 public class ConsignmentItem {
 
-  @Schema(description = "The associated booking number provided by the carrier for this `Consignment Item`.", example = "ABC709951", maxLength = 35, pattern = "^\\S(?:.*\\S)?$")
-  private String carrierBookingReference;
+  public static final String CLASS_SCHEMA_DESCRIPTION = "Defines a list of `CargoItems` belonging together and the associated `Booking`. A `ConsignmentItem` can be split across multiple containers by referencing multiple `CargoItems`.";
+
+  @Schema(requiredMode = Schema.RequiredMode.REQUIRED, description = "The associated booking number provided by the carrier for this `Consignment Item`.", example = "ABC709951", maxLength = 35, pattern = "^\\S(?:.*\\S)?$")
+  protected String carrierBookingReference;
 
   @Schema(
+      requiredMode = Schema.RequiredMode.REQUIRED,
       description =
 """
 A plain language description that is precise enough for Customs services to be able to identify the goods. General terms (i.e. 'consolidated', 'general cargo' 'parts' or 'freight of all kinds') or not sufficiently precise description cannot be accepted.
 
 **Condition:** The order of the items in this array **MUST** be preserved as by the provider of the API.
 """)
-  @ArraySchema(
-      schema =
-          @Schema(
-              description = "A line describing the cargo",
-              example = "blue shoes size 47",
-              maxLength = 35,
-              pattern = "^\\S(?:.*\\S)?$"))
-  private List<String> descriptionOfGoods;
+  @ArraySchema(maxItems = 150)
+  protected List<DescriptionOfGoodsLine> descriptionOfGoods;
 
-  @Schema(name = "HSCodes", description = "A list of `HS Codes` that apply to this `consignmentItem`")
-  @ArraySchema(schema = @Schema(description = "Used by customs to classify the product. Must be 6 to 10 digits.", example = "851713", minLength = 6, maxLength = 10, pattern = "^\\d{6,10}$"), minItems = 1)
-  private List<String> hsCodes;
+  @Schema(requiredMode = Schema.RequiredMode.REQUIRED, name = "HSCodes", description = "A list of `HS Codes` that apply to this `consignmentItem`")
+  @ArraySchema(minItems = 1)
+  protected List<HSCode> hsCodes;
 
   @Schema(description = "A list of `National Commodity Codes` that apply to this `commodity`")
-  private List<NationalCommodityCode> nationalCommodityCodes;
+  protected List<NationalCommodityCode> nationalCommodityCodes;
 
   @Schema(
       description =
@@ -42,27 +42,22 @@ A list of the `ShippingMarks` applicable to this `consignmentItem`
 
 **Condition:** The order of the items in this array **MUST** be preserved as by the provider of the API.
 """)
-  @ArraySchema(
-      schema =
-          @Schema(
-              description = "Identifying details or markings on the packages.",
-              example = "Made in China",
-              maxLength = 35),
-      maxItems = 50)
-  private List<String> shippingMarks;
+  @ArraySchema(maxItems = 50)
+  protected List<ShippingMark> shippingMarks;
 
-  @Schema(description = "A list of all `cargoItems`")
-  private List<CargoItem> cargoItems;
+  @Schema(requiredMode = Schema.RequiredMode.REQUIRED, description = "A list of all `cargoItems`")
+  @ArraySchema(minItems = 1)
+  protected List<CargoItem> cargoItems;
 
   @Schema(description = "Export license for the consignment item.")
-  private ExportLicense exportLicense;
+  protected ExportLicense exportLicense;
 
   @Schema(description = "Import license for the consignment item.")
-  private ImportLicense importLicense;
+  protected ImportLicense importLicense;
 
   @Schema(description = "A list of `References`")
-  private List<ReferenceConsignmentItem> references;
+  protected List<ReferenceConsignmentItem> references;
 
   @Schema(description = "A list of `Customs references`")
-  private List<CustomsReference> customsReferences;
+  protected List<CustomsReference> customsReferences;
 }