+++ specifications/src/main/java/org/dcsa/conformance/specifications/standards/dt/v100/model/Reference.java
@@ -3,11 +3,15 @@ package org.dcsa.conformance.specifications.standards.dt.v100.model;
 import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.Data;
 
-@Schema(description = "References provided by the shipper or freight forwarder at the time of Booking or Shipping Instructions. Carriers share them back during Track & Trace and they may appear on the Bill of Lading.")
+@Schema(description = Reference.CLASS_SCHEMA_DESCRIPTION)
 @Data
 public class Reference {
 
+  public static final String CLASS_SCHEMA_DESCRIPTION =
+      "References provided by the shipper or freight forwarder at the time of Booking or Shipping Instructions. Carriers share them back during Track & Trace and they may appear on the Bill of Lading.";
+
   @Schema(
+      requiredMode = Schema.RequiredMode.REQUIRED,
       description =
 """
 The reference type codes defined by DCSA. Possible values are:
@@ -16,8 +20,13 @@ The reference type codes defined by DCSA. Possible values are:
 """,
       example = "CR",
       maxLength = 3)
-  private String type;
+  protected String type;
 
-  @Schema(description = "The value of the reference.", example = "HHL00103004", maxLength = 35, pattern = "^\\S(?:.*\\S)?$")
-  private String value;
+  @Schema(
+      requiredMode = Schema.RequiredMode.REQUIRED,
+      description = "The value of the reference.",
+      example = "HHL00103004",
+      maxLength = 35,
+      pattern = "^\\S(?:.*\\S)?$")
+  protected String value;
 }