+++ specifications/src/main/java/org/dcsa/conformance/specifications/standards/ebl/v300/model/PartyAddress.java
@@ -7,7 +7,7 @@ import lombok.Data;
 @Data
 public class PartyAddress {
 
-  @Schema(description = "The name of the street of the party’s address.", example = "Ruijggoordweg", maxLength = 70)
+  @Schema(requiredMode = Schema.RequiredMode.REQUIRED, description = "The name of the street of the party’s address.", example = "Ruijggoordweg", maxLength = 70)
   private String street;
 
   @Schema(description = "The number of the street of the party’s address.", example = "100", maxLength = 50)
@@ -22,7 +22,7 @@ public class PartyAddress {
   @Schema(name = "POBox", description = "A numbered box at a post office where a person or business can have mail or parcels delivered.", example = "123", maxLength = 20)
   private String poBox;
 
-  @Schema(description = "The city name of the party’s address.", example = "Amsterdam", maxLength = 35, pattern = "^\\S(?:.*\\S)?$")
+  @Schema(requiredMode = Schema.RequiredMode.REQUIRED, description = "The city name of the party’s address.", example = "Amsterdam", maxLength = 35, pattern = "^\\S(?:.*\\S)?$")
   private String city;
 
   @Schema(
@@ -43,6 +43,6 @@ More info can be found here: [UN/LOCODE](https://unece.org/trade/cefact/UNLOCODE
   @Schema(description = "The state/region of the party’s address.", example = "North Holland", maxLength = 65)
   private String stateRegion;
 
-  @Schema(description = "The 2 characters for the country code using [ISO 3166-1 alpha-2](https://www.iso.org/obp/ui/#iso:pub:PUB500001:en)", example = "NL", minLength = 2, maxLength = 2, pattern = "^[A-Z]{2}$")
+  @Schema(requiredMode = Schema.RequiredMode.REQUIRED, description = "The 2 characters for the country code using [ISO 3166-1 alpha-2](https://www.iso.org/obp/ui/#iso:pub:PUB500001:en)", example = "NL", minLength = 2, maxLength = 2, pattern = "^[A-Z]{2}$")
   private String countryCode;
 }